Revert "Don't test the engine on startup"
This reverts commit ca0dae9dbf18342dedf42f6e07447cca4daf11d3.

This is needed because we load the engine to get function pointers for
get_item_info and item_set_cas, then destroy it, which can segfault if
we haven't initialized the engine.

The worst problem is that we create a db in /tmp; we can avoid this by
not loading a database if no dbname is given.

Change-Id: I45deb3dcfa37056e8fd639ef0aaab5b302c91ca0
Reviewed-on: http://review.northscale.com:8080/408
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>