bug 2182 - Properly handle engine-specific data.
Connection specific "engine-specific" storage ended up being tied to
the instance of the inferior engine and not related to the cookie.

This change ensures the inferior engine-specific data and
bucket_engine engine-specific data are handled properly by tracking
them with a separate chunk of per-connection memory.

Refactoring was required to ensure that all access to this memory went
through the correct paths.

Change-Id: I93864cc1cca9b21f9ee0febaf1fde5495d7f8f55
Reviewed-on: http://review.northscale.com/2085
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>