Started to add documentation to the functions
I've tried to walk through the file and add a brief textual description
of the functions there. I've added some notes to myself about stuff
that should be looked into, because I didn't want to modify code in
the same commit with the documentation. These comments will be addressed
in subsequent commits and then removed..

Change-Id: Ie60b45822eac69ce568754e96d40938fafb0213d
Reviewed-on: http://review.membase.org/6472
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>