Use a special test to get the engine handle for the bastard functions
Parts of the function in the engine specification doesn't work
very well with bucket engine and the fact that the underlying engine
may disappear at any time (functions like item_set_cas etc).

Let's just try to perform them if it is possible, otherwise we'll
just "ignore them", and await for the client to perform the next
real operation through the engine interface to disconnect the
connection.

Change-Id: Ibaa227404a4220e1f970254e6cf8ac885a5ab056
Reviewed-on: http://review.membase.org/6493
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>