Lock free in normal operations.
The concept is pretty simple.  *Most* of the time, the connection
contains a pointer to its current engine and can just inspect it when
it needs to.  If the state is "running", it's free to do whatever it
needs to the engine.  If the state is not running, it's shutting down
and the connection should begin this procedure.

In creation, location and destruction, there's a lock protecting
larger shared state in the server proceeding through these sensitive
areas.

Change-Id: Id2852bc99c2e1c0f297f9737cc7c47b12357e00d
Reviewed-on: http://review.couchbase.org/13812
Tested-by: Jin Lim <jin@couchbase.com>
Reviewed-by: Jin Lim <jin@couchbase.com>