fixed get_engine_handle to never return STATE_STOPPED handle
Instead of checking state and then incrementing clients count we
increment clients count and then check.

So when we check clients count is > 0 and nobody can transition from
STATE_STOPPING to STATE_STOPPED. The only transition that can happen
is from running to stopping but that's fine. If we see peh in state
not running we decrement clients count and return 0.

We can also see that returned engine handle not only cannot be in
STATE_STOPPED but cannot transition into this state until
release_engine_handle is called.

We also don't lock engines mutex. There's no need. We just need to
check that e->clients == 0 was reached. And because we're not in
STATE_RUNNING some clients increments can only be temporary. It is
impossible for client to increment clients and observe STATE_RUNNING
when maybe_start_engine_shutdown observed clients == 0 and
STATE_STOPPING.

Change-Id: Id99c129d78dd5355e3ebff44c3331336387e19fb
Reviewed-on: http://review.couchbase.org/13869
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>