MB-5081 Correct the assertion predicate on reserve_cookie()
The number of clients per bucket is incremented whenever a
memcached worker thread visits the downstream engine and then
decremented upon leaving from the downstream engine.

reserve_cookie() function is mainly invoked when a memcached
worker thread invokes the downstream engine's TAP connection
API, which means that the number of clients on that bucket is
greater than zero at the moment of API invocation.

As reserve_cookie() function is also invoked from the separate
thread in the downstream engine whenever a connection cookie is
notified to the upstream engine, "number of clients" predicate
on the assertion check should be greater than equal to zero.

Change-Id: Idc20223712ca97654d69fb128751bdc862f7c7cb
Reviewed-on: http://review.couchbase.org/15379
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>